src-base: add package
src-base refers to vendor/XXXXX/base to contain the dependent
file for router complation.  asuswrt used its Maekfile to trigger
the full router compilation. Several files like version.h are
generated within the process.

This package refer to the source code but doesn't follow the
entire process. It generates a fake Makefile.asus only to trigger
the needed target to generate the mentioned files, and doesn't build
the router actually. The package "router" would accomplish the
compilation for router and its components under the directory.

Change-Id: I05464ffe9a7c1e3e8b1237d71e5766ae44fad85f